Fix when captype modules update results_cap table
We need to import the capacity_all results first before updating the
table; otherwise, the update statement has no effect.
